# baseline-db/openapi.yml
openapi: 3.0.0
info:
  title: merkle-tree
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:9000
    variables: {}
paths:
...
  /node:
    post:
      tags:
        - node
      summary: /node -> inserts a node object
      description: inserts a node object
      operationId: /node->insertsanodeobject
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1node-%3Einsertsanodeobjectrequest'
            example:
              contractName: MerkleTreeControllerSHA
              node:
                value: 0xabc123-0
                nodeIndex: 0
                isLocked: false
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /nodes:
    patch:
      tags:
        - node
      summary: /nodes -> updates many node objects
      description: inserts a leaf object
      operationId: /nodes->updatesmanynodeobjects
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1nodes-%3Eupdatesmanynodeobjectsrequest'
            example:
              contractName: MerkleTreeControllerSHA
              nodes:
                - value: 0xabc123-0
                  nodeIndex: 0
                - value: 0xabc123-1
                  nodeIndex: 1
                - value: 0xabc123-2
                  nodeIndex: 2
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - node
      summary: /nodes -> get nodes by array of indices
      operationId: /nodes->getnodesbyarrayofindices
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...
  /node/index:
    get:
      tags:
        - node
      summary: /node/index -> gets a node at nodeIndex
      operationId: /node/index->getsanodeatnodeIndex
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /node/value:
    get:
      tags:
        - node
      summary: /node/value -> gets a node with a given value
      operationId: /node/value->getsanodewithagivenvalue
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /nodes/count:
    get:
      tags:
        - node
      summary: /nodes/count -> count the number of nodes in the db
      operationId: /nodes/count->countthenumberofnodesinthedb
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata/contractAddress:
    post:
      tags:
        - metaData
      summary: /metadata/contractAddress -> inserts the contract address of the tree
      description: inserts a node object
      operationId: /metadata/contractAddress->insertsthecontractaddressofthetree
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1metadata~1contractAddress-%3EinsertsthecontractaddressofthetreeRequest'
            example:
              contractName: MerkleTreeControllerSHA
              contractAddress: 0xabc123-0
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - metaData
      summary: /metadata/contractAddress -> gets the contractAddress
      description: inserts a leaf object
      operationId: /metadata/contractAddress->getsthecontractAddress
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata/contractInterface:
    post:
      tags:
        - metaData
      summary: /metadata/contractInterface -> inserts the contract address of the tree
      description: inserts a node object
      operationId: /metadata/contractInterface->insertsthecontractaddressofthetree
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: string
              example: >-
                {
                	"contractName": "MerkleTreeControllerSHA",
                	"contractInterface": "{ "pretend": "this is a stringified json interface" }"
                }
            example: >-
              {
              	"contractName": "MerkleTreeControllerSHA",
              	"contractInterface": "{ "pretend": "this is a stringified json interface" }"
              }
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - metaData
      summary: /metadata/contractInterface -> gets the contractInterface
      description: inserts a leaf object
      operationId: /metadata/contractInterface->getsthecontractInterface
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata/treeHeight:
    post:
      tags:
        - metaData
      summary: /metadata/treeHeight -> establishes the treeHeight for a db
      description: inserts a node object
      operationId: /metadata/treeHeight->establishesthetreeHeightforadb
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1metadata~1treeHeight-%3EestablishesthetreeHeightforadbRequest'
            example:
              contractName: MerkleTreeControllerSHA
              treeHeight: 22
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata:
    get:
      tags:
        - metaData
      summary: /metadata -> gets the tree's metadata object
      operationId: /metadata->getsthetree'smetadataobject
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata/latestLeaf:
    patch:
      tags:
        - metaData
      summary: /metadata/latestLeaf -> updates the latestLeaf metadata object
      description: inserts a leaf object
      operationId: /metadata/latestLeaf->updatesthelatestLeafmetadataobject
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1metadata~1latestLeaf-%3EupdatesthelatestLeafmetadataobjectRequest'
            example:
              contractName: MerkleTreeControllerSHA
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - metaData
      summary: /metadata/latestLeaf -> get the latestLeaf metadata object
      description: inserts a leaf object
      operationId: /metadata/latestLeaf->getthelatestLeafmetadataobject
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
  /metadata/latestRecalculation:
    patch:
      tags:
        - metaData
      summary: /metadata/latestRecalculation -> updates the latestRecalculation metadata object
      description: inserts a leaf object
      operationId: /metadata/latestRecalculation->updatesthelatestRecalculationmetadataobject
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/~1metadata~1latestRecalculation-%3EupdatesthelatestRecalculationmetadataobjectRequest'
            example:
              contractName: MerkleTreeControllerSHA
        required: true
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
        - metaData
      summary: /metadata/latestRecalculation -> gets the latestRecalculation metadata object
      description: inserts a leaf object
      operationId: /metadata/latestRecalculation->getsthelatestRecalculationmetadataobject
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          style: simple
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ''
          headers: {}
      deprecated: false
...